refactor: move Cards to component
