+++ app/src/main/java/kr/ksw/visitkorea/presentation/home/component/CultureCard.kt
@@ -1,13 +1,14 @@
-package kr.ksw.visitkorea.presentation.home.screen
+package kr.ksw.visitkorea.presentation.home.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.LocationOn
@@ -15,11 +16,9 @@ import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
@@ -34,29 +33,26 @@ import kr.ksw.visitkorea.presentation.component.SingleLineText
 import kr.ksw.visitkorea.presentation.ui.theme.VisitKoreaTheme
 
 @Composable
-fun RestaurantCard(
+fun CultureCard(
+    modifier: Modifier = Modifier,
+    ratio: Float = 0.7f,
     title: String,
     address: String,
-    dist: String,
-    category: String,
     image: String,
 ) {
-    Card(
-        modifier = Modifier
-            .width(300.dp),
-        elevation = CardDefaults.elevatedCardElevation(6.dp),
+    Card (
+        modifier = modifier
+            .aspectRatio(ratio),
+        shape = RoundedCornerShape(24.dp),
+        elevation = CardDefaults.elevatedCardElevation(
+            defaultElevation = 4.dp
+        )
     ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.White)
-                .padding(10.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
+        Column {
             AsyncImage(
                 modifier = Modifier
-                    .size(88.dp)
-                    .clip(RoundedCornerShape(16.dp))
+                    .fillMaxWidth()
+                    .aspectRatio(1f)
                     .background(color = Color.LightGray),
                 model = ImageRequest
                     .Builder(LocalContext.current)
@@ -68,23 +64,17 @@ fun RestaurantCard(
             )
             Column(
                 modifier = Modifier
-                    .padding(start = 8.dp)
+                    .fillMaxSize()
+                    .background(Color.White)
+                    .padding(horizontal = 8.dp),
+                verticalArrangement = Arrangement.Center
             ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween
-                ) {
-                    SingleLineText(
-                        text = title,
-                        fontSize = 16.sp,
-                        fontWeight = FontWeight.Medium
-                    )
-                    Text(
-                        text = category,
-                        fontSize = 14.sp,
-                        color = Color.Gray
-                    )
-                }
+                SingleLineText(
+                    modifier = Modifier.padding(start = 2.dp),
+                    text = title,
+                    fontSize = 14.sp,
+                    fontWeight = FontWeight.Medium
+                )
                 Row(
                     verticalAlignment = Alignment.CenterVertically
                 ) {
@@ -96,13 +86,9 @@ fun RestaurantCard(
                     )
                     SingleLineText(
                         text = address,
-                        fontSize = 14.sp
+                        fontSize = 12.sp,
                     )
                 }
-                SingleLineText(
-                    text = dist,
-                    fontSize = 12.sp
-                )
             }
         }
     }
@@ -110,15 +96,16 @@ fun RestaurantCard(
 
 @Composable
 @Preview(showBackground = true)
-fun RestaurantCardPreview() {
+fun CultureCardPreview() {
     VisitKoreaTheme {
-        Surface {
-            RestaurantCard(
-                "음식점",
-                "음식점 주소",
-                "188m",
-                "한식",
-                "https"
+        Surface(
+            modifier = Modifier
+                .padding(20.dp)
+        ) {
+            CultureCard(
+                title = "문화시설",
+                address = "문화시설 주소",
+                image = "https://ksw"
             )
         }
     }
